% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function3.R
\name{cortable}
\alias{cortable}
\title{Function 3}
\usage{
cortable(YearInput3)
}
\arguments{
\item{year}{Input should be a year.}
}
\description{
This function allows you to get correlation matrix of all the numerical variables we have in the dataset in a particular year.
}
\examples{
cortable(1950)
}
\keyword{cortable}
