% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2.R
\name{OldestPlyr}
\alias{OldestPlyr}
\title{Function 2}
\usage{
OldestPlyr(YearInput2)
}
\arguments{
\item{year}{Input should be a year.}
}
\description{
This function allows you to know the oldest player in the year you put in.
}
\examples{
OldestPlyr(1950)
}
\keyword{OldestPlyr}
