% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{TmPlayer}
\alias{TmPlayer}
\title{Function 1}
\usage{
TmPlayer(YearInput)
}
\arguments{
\item{year}{Input should be a year.}
}
\description{
This function allows you to know a player and his team.
}
\examples{
TmPlayer(1950)
}
\keyword{TmPlayer}
